renamed connectors to database_connectors and the base Connector to DatabaseConnector
