--- ai_text_to_sql/connectors/__init__.py
@@ -1,4 +1,0 @@
-from ai_text_to_sql.connectors.sqlite_connector import SQLiteConnector
-from ai_text_to_sql.connectors.mssql_connector import MSSQLConnector
-from ai_text_to_sql.connectors.mysql_connector import MySQLConnector
-from ai_text_to_sql.connectors.postgresql_connector import PostgreSQLConnector
