+++ ai_text_to_sql/database_connectors/connector_factory.py
@@ -2,16 +2,16 @@ import sys
 import inspect
 from typing import Optional, Dict, Type
 
-from ai_text_to_sql.connectors import *
-from .connector import Connector
+from ai_text_to_sql.database_connectors import *
+from .database_connector import DatabaseConnector
 
 
 class ConnectorFactory:
     """
-    The class for building database connectors.
+    The class for building database database_connectors.
     """
     @staticmethod
-    def build_connector(name: str, connection_data: Optional[Dict]) -> Connector:
+    def build_connector(name: str, connection_data: Optional[Dict]) -> DatabaseConnector:
         """
         Build a database connector.
         :param name: The name of the connector.
@@ -25,13 +25,13 @@ class ConnectorFactory:
             raise ValueError(f"Unsupported connector: {name}")
 
     @staticmethod
-    def _discover_connectors() -> Dict[str, Type[Connector]]:
+    def _discover_connectors() -> Dict[str, Type[DatabaseConnector]]:
         """
-        Discover available connectors dynamically.
+        Discover available database_connectors dynamically.
         :return: A dictionary mapping connector names to their corresponding classes.
         """
         connectors = {}
         for name, obj in inspect.getmembers(sys.modules[__name__]):
-            if inspect.isclass(obj) and issubclass(obj, Connector) and obj is not Connector:
+            if inspect.isclass(obj) and issubclass(obj, DatabaseConnector) and obj is not DatabaseConnector:
                 connectors[obj.__name__[:-9]] = obj  # Remove "Connector" from class name
         return connectors
